+++ src/parser.ts
@@ -1,30 +1,5 @@
-import {
-  possibly,
-  anythingExcept,
-  many,
-  choice,
-  char,
-  between,
-  str,
-  coroutine,
-  sequenceOf,
-  letters,
-  regex
-} from './lib/arcsecond';
-
-// types
-enum TokenType {
-  selfClosingTag = 'self-closing-tag',
-  openTag = 'open-tag',
-  closeTag = 'close-tag',
-  text = 'text'
-}
-
-type Token = {
-  type: TokenType;
-  ignore?: boolean;
-  value: string;
-};
+import { Scanner } from './Scanner';
+import { LexemeTag, LexemeType } from './types';
 
 export type TextElement = {
   type: 'text';
@@ -39,52 +14,7 @@ export type TagElement = {
 
 export type ParserElement = TextElement | TagElement;
 
-// tagger
-const tokenTag = (
-  type: TokenType,
-  customizer = (id: ParserElement[] | null) => id as any
-) => (value: ParserElement[] | null) => ({
-  type,
-  value: customizer(value)
-});
-
-// parser
-const lAngle = char('<');
-const rAngle = char('>');
-const lettersOrDigits = regex(/^[a-zA-Z0-9]+/);
-
-const openTag = between(lAngle)(rAngle)(lettersOrDigits).map(
-  tokenTag(TokenType.openTag)
-);
-
-const closeTag = between(str('</'))(rAngle)(lettersOrDigits).map(
-  tokenTag(TokenType.closeTag)
-);
-
-const selfClosingTag = between(lAngle)(str('/>'))(
-  sequenceOf([lettersOrDigits, possibly(char(' '))])
-).map(tokenTag(TokenType.selfClosingTag));
-
-const text = many(
-  anythingExcept(choice([openTag, closeTag, selfClosingTag]))
-).map(tokenTag(TokenType.text, val => (val ? val.join('') : '')));
-
-// order matters
-const token = choice([selfClosingTag, openTag, closeTag, text]);
-
-const tokenize = coroutine(function* () {
-  const result = [];
-
-  while (true) {
-    const value = yield token;
-    if (value.value === '' || value.isError) break;
-    result.push(value);
-  }
-
-  return result;
-});
-
-const structBuilder = (tokens: Token[], openTags = [] as string[]) => {
+const structBuilder = (tokens: LexemeTag[], openTags = [] as string[]) => {
   const result = [] as ParserElement[];
 
   for (let i = 0; i < tokens.length; i++) {
@@ -94,49 +24,48 @@ const structBuilder = (tokens: Token[], openTags = [] as string[]) => {
     token.ignore = true;
 
     switch (token.type) {
-      case TokenType.selfClosingTag: {
+      case LexemeType.HTML_SELFCLOSING_TAG: {
         result.push({
           type: 'tag',
           // [ 'br', null ] tagtype and possible space or null
-          tagType: token.value[0],
+          tagType: token.name,
           value: null
         });
 
         break;
       }
-      case TokenType.text: {
-        const { ignore, ...textObj } = token;
-        result.push(textObj as any);
+      case LexemeType.STRING: {
+        result.push({ type: 'text', value: token.value });
         break;
       }
-      case TokenType.openTag: {
+      case LexemeType.HTML_OPENING_TAG: {
         result.push({
           type: 'tag',
-          tagType: token.value,
-          value: structBuilder(tokens.slice(i + 1), [...openTags, token.value])
+          tagType: token.name,
+          value: structBuilder(tokens.slice(i + 1), [...openTags, token.name])
         });
 
         break;
       }
-      case TokenType.closeTag: {
+      case LexemeType.HTML_CLOSING_TAG: {
         const lastOpen = openTags[openTags.length - 1];
-        if (lastOpen === token.value) {
+        if (lastOpen === token.name) {
           return result;
         }
 
-        throw new Error(`Expected ${lastOpen} but got ${token.value} instead.`);
+        throw new Error(`Expected ${lastOpen} but got ${token.name} instead.`);
       }
       default:
-        throw new Error(`Unknown type ${token.type}`);
+        throw new Error(`Unknown type ${token}`);
     }
   }
   return result;
 };
 
 const parse = (text: string) => {
-  const tokens = tokenize.run(text).result;
+  const scanner = new Scanner(text);
 
-  return structBuilder(tokens);
+  return structBuilder(scanner.scanTokensWithMerge());
 };
 
 export { parse };
